01_YOLOv4_VDA_4994_DATASET_GENERATOR/requirements.txt
Augmentor
absl-py		
apipkg	
asgiref		
asn1crypto		
astor	
astroid		
atomicwrites
attrs	
awscli		
backcall		
beautifulsoup4		
blas		
bleach	
botocore	
bottleneck		
ca-certificates		
certifi		
cffi		
chardet		
chart-studio	
cloudpickle	
cmake	
colorama	
colorlover	
contextlib2	
cryptography	
cudatoolkit	
cufflinks	
cvxopt	
cycler	
cymem	
cython-blis	
cytoolz	
dask-core	
dataclasses	
decorator	
defusedxml	
django	
docutils	
entrypoints	
execnet	
fastai	
fastprogress	
flake8	
flake8-polyfill	
flake8-quotes	
freetype	
gast	
glpk	
google-images-download	
google-pasta	
grpcio	
gsl	
h5py	
hdf5	
html5lib	
icc_rt	
icu	
idna	
image	
imageio	
importlib_metadata	
imutils	
intel-openmp	
ipykernel	
ipython	
ipython_genutils	
ipywidgets	
isort	
jedi	
jinja2	
jmespath	
joblib	
jpeg	
json	
json5	
jsonschema	
jupyter	
jupyter-console	
jupyter_client	
jupyter_core	
jupyterlab	
jupyterlab_server	
kaggle	
keras	
keras-applications	
keras-base	
keras-metrics	
keras-preprocessing	
kiwisolver	
lazy-object-proxy	
libgpuarray	
libiconv	
libopencv	
libpng	
libprotobuf	
libpython	
libsodium	
libtiff	
libxml2	
libxslt	
lxml	
m2w64-binutils	
m2w64-bzip2	
m2w64-crt-git	
m2w64-gcc	
m2w64-gcc-ada	
m2w64-gcc-fortran	
m2w64-gcc-libgfortran	
m2w64-gcc-libs	
m2w64-gcc-libs-core	
m2w64-gcc-objc	
m2w64-gmp	
m2w64-headers-git	
m2w64-isl	
m2w64-libiconv	
m2w64-libmangle-git	
m2w64-libwinpthread-git	
m2w64-make	
m2w64-mpc	
m2w64-mpfr	
m2w64-pkg-config	
m2w64-toolchain	
m2w64-tools-git	
m2w64-windows-default-manifest	
m2w64-winpthreads-git	
m2w64-zlib	
mako	
markdown	
markupsafe	
matplotlib	
mccabe	
mistune	
mkl	
mkl-service	
mkl_fft	
mkl_random	
mock	
more-itertools	
msys2-conda-epoch	
murmurhash	
nbconvert	
nbformat	
networkx	
ninja	
notebook	
numexpr	
numpy	
numpy-base	
nvidia-ml-py3	
olefile	
opencv	
opencv-contrib-python	
opencv-python	
openssl	
opt-einsum	
packaging	
pandas	
pandoc	
pandocfilters	
parso	
path	
path.py	
patsy	
pep8-naming	
pickleshare	
pillow	
pip	
plac	
plotly	
pluggy	
preshed	
prometheus_client	
prompt_toolkit	
protobuf	
py	
py-opencv	
pyasn1	
pycodestyle	
pycparser	
pyflakes	
pygments	
pygpu	
pylint	
pyopenssl	
pyparsing	
pyqt	
pyqt5-sip	
pyrsistent	
pysocks	
pytesseract	
pytest	
pytest-shutil	
python	
python-dateutil	
python-slugify	
pytorch	
pytz	
pywavelets	
pywin32	
pywinpty	
pyyaml	
pyzmq	
qt	
qtconsole	
qutil	
requests	
retrying	
rsa	
s3transfer	
scikit-image	
scikit-learn	
scipy	
seaborn	
selenium	
send2trash	
setuptools	
simplejson	
sip	
six	
soupsieve	
spacy	
sqlalchemy	
sqlite	
sqlparse	
srsly	
statsmodels	
tensorboard	
tensorflow	
tensorflow-estimator	
tensorflow-gpu	
termcolor	
terminado	
testpath	
text-unidecode	
theano	
thinc	
tk	
toml	
toolz	
torchvision	
tornado	
tqdm	
traitlets	
unidecode	
urllib3	
vc	
vs2015_runtime	
wasabi	
wcwidth	
webencodings	
werkzeug	
wheel	
widgetsnbextension	
win_inet_pton	
wincertstore	
winpty	
wrapt	
xz	
yaml	
zeromq	
zipp	
zlib	